# cliente-promociones/api/openapi.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- description: Generated server url
  url: http://localhost:8080
tags:
- description: Operaciones relacionadas con promociones
  name: Promociones
paths:
  /api/promociones/{id}:
    delete:
      description: Elimina una promoción específica dado su ID
      operationId: eliminarPromocion
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          description: OK
      summary: Eliminar una promoción
      tags:
      - Promociones
      x-accepts: application/json
    get:
      description: Devuelve los detalles de una promoción específica dado su ID
      operationId: obtenerPromocionPorId
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PromocionEntity'
          description: OK
      summary: Obtener una promoción por ID
      tags:
      - Promociones
      x-accepts: '*/*'
    put:
      description: Actualiza los datos de una promoción ya existente utilizando su
        ID
      operationId: actualizarPromocion
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PromocionEntity'
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PromocionEntity'
          description: OK
      summary: Actualizar una promoción existente
      tags:
      - Promociones
      x-content-type: application/json
      x-accepts: '*/*'
  /api/promociones:
    get:
      description: Devuelve una lista con todas las promociones disponibles
      operationId: obtenerPromociones
      responses:
        "200":
          content:
            '*/*':
              schema:
                items:
                  $ref: '#/components/schemas/PromocionEntity'
                type: array
          description: OK
      summary: Obtener todas las promociones
      tags:
      - Promociones
      x-accepts: '*/*'
    post:
      description: Crea una nueva promoción si no existe una con el mismo ID
      operationId: crearPromocion
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PromocionEntity'
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PromocionEntity'
          description: OK
      summary: Crear una nueva promoción
      tags:
      - Promociones
      x-content-type: application/json
      x-accepts: '*/*'
  /api/promociones/aplicar:
    post:
      description: Aplica las promociones disponibles a un ticket dado y devuelve
        el ticket actualizado
      operationId: aplicarPromociones
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Ticket'
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Ticket'
          description: OK
      summary: Aplicar promociones a un ticket
      tags:
      - Promociones
      x-content-type: application/json
      x-accepts: '*/*'
components:
  schemas:
    PromocionEntity:
      description: Entidad que representa una promoción aplicable a un artículo
      example:
        descripcion: 20% de descuento en el producto A
        tipo: DESCUENTO
        fechaAlta: 2024-04-01T00:00:00.000+00:00
        fechaInicio: 2024-04-15T00:00:00.000+00:00
        porcentajeDescuento: 20
        id: PROMO2024
        fechaFin: 2024-05-15T00:00:00.000+00:00
        precioPromocion: 5.99
        codigoArticulo: A12345
      properties:
        id:
          description: ID único de la promoción
          example: PROMO2024
          type: string
        descripcion:
          description: Descripción de la promoción
          example: 20% de descuento en el producto A
          type: string
        fechaAlta:
          description: Fecha de alta de la promoción
          example: 2024-04-01
          format: date
          type: string
        fechaInicio:
          description: Fecha de inicio de la promoción
          example: 2024-04-15
          format: date
          type: string
        fechaFin:
          description: Fecha de finalización de la promoción
          example: 2024-05-15
          format: date
          type: string
        codigoArticulo:
          description: Código del artículo al que se aplica la promoción
          example: A12345
          type: string
        tipo:
          description: Tipo de promoción (PRECIO o DESCUENTO)
          enum:
          - PRECIO
          - DESCUENTO
          example: DESCUENTO
          type: string
        precioPromocion:
          description: Nuevo precio promocional (solo si el tipo es PRECIO)
          example: 5.99
          type: number
        porcentajeDescuento:
          description: Porcentaje de descuento (solo si el tipo es DESCUENTO)
          example: 20
          type: number
      required:
      - codigoArticulo
      - fechaFin
      - fechaInicio
      - id
      - tipo
      type: object
    LineaVenta:
      description: Representa una línea de venta dentro de un ticket
      example:
        idPromocionAplicada: PROMO2024
        precioUnitarioOriginal: 10.99
        importeTotal: 17.98
        precioUnitarioPromocionado: 8.99
        cantidad: 2
        codigoArticulo: A12345
      properties:
        codigoArticulo:
          description: Código del artículo vendido
          example: A12345
          type: string
        cantidad:
          description: Cantidad de artículos vendidos
          example: 2
          format: int32
          minimum: 1
          type: integer
        precioUnitarioOriginal:
          description: Precio original por unidad
          example: 10.99
          type: number
        precioUnitarioPromocionado:
          description: Precio por unidad después de aplicar promociones
          example: 8.99
          type: number
        importeTotal:
          description: Importe total de la línea (precio promocionado x cantidad)
          example: 17.98
          type: number
        idPromocionAplicada:
          description: ID de la promoción aplicada (si existe)
          example: PROMO2024
          type: string
      required:
      - codigoArticulo
      type: object
    Ticket:
      description: Representa un ticket de compra que contiene varias líneas de venta
      example:
        lineasVenta:
        - idPromocionAplicada: PROMO2024
          precioUnitarioOriginal: 10.99
          importeTotal: 17.98
          precioUnitarioPromocionado: 8.99
          cantidad: 2
          codigoArticulo: A12345
        - idPromocionAplicada: PROMO2024
          precioUnitarioOriginal: 10.99
          importeTotal: 17.98
          precioUnitarioPromocionado: 8.99
          cantidad: 2
          codigoArticulo: A12345
        - idPromocionAplicada: PROMO2024
          precioUnitarioOriginal: 10.99
          importeTotal: 17.98
          precioUnitarioPromocionado: 8.99
          cantidad: 2
          codigoArticulo: A12345
        - idPromocionAplicada: PROMO2024
          precioUnitarioOriginal: 10.99
          importeTotal: 17.98
          precioUnitarioPromocionado: 8.99
          cantidad: 2
          codigoArticulo: A12345
        - idPromocionAplicada: PROMO2024
          precioUnitarioOriginal: 10.99
          importeTotal: 17.98
          precioUnitarioPromocionado: 8.99
          cantidad: 2
          codigoArticulo: A12345
        id: TCK123456
      properties:
        id:
          description: ID único del ticket
          example: TCK123456
          type: string
        lineasVenta:
          description: Listado de líneas de venta incluidas en el ticket
          items:
            $ref: '#/components/schemas/LineaVenta'
          maxItems: 2147483647
          minItems: 1
          type: array
      required:
      - id
      type: object